+++ app/javascript/bundles/TwitterTrends/components/ShowTrend.jsx
@@ -6,7 +6,7 @@ import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
 import Typography from '@material-ui/core/Typography';
 
-import ShowTweet from './ShowTweet';
+import ShowTweet from './ShowTweets';
 
 const styles = theme => ({});
